refactor: add file pages.ts
