+++ src/pages/pages.ts
@@ -1,0 +1,22 @@
+export const pages = [
+    {
+        name: "Tabela Nutricional",
+        link: "/tabela-nutricional"
+    },
+    {
+        name: "Benefícios",
+        link: "/beneficios"
+    },
+    {
+        name: "Como Tomar",
+        link: "/como-tomar"
+    },
+    {
+        name: "Ingredientes",
+        link: "/ingredientes"
+    },
+    // {
+    //     name: "Sobre",
+    //     link: "/descricao"
+    // },
+];
